refactor(components): colocate stories alongside implementations
Move 6 story files from top-level stories/ directory into their
respective component directories. Update Storybook config to remove
the old stories/ glob.
