+++ packages/components/src/components/Header/Header.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react-webpack5';
-import { Header } from '../src/components/Header';
+import { Header } from '../Header';
 
 const meta = {
   title: 'Header',
