+++ packages/components/src/components/JsonForm/JsonForm.stories.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import type { Meta, StoryObj } from '@storybook/react-webpack5';
-import { JsonForm } from '../src/components/JsonForm/JsonForm';
+import { JsonForm } from './JsonForm';
 
 const meta: any = {
   title: 'Json Form',
